Reduced dupliacte codes in linked stack and linked queue. (#19)
