+++ data_structures/sequence/stack_queue/linked_queue.py
@@ -5,9 +5,9 @@ doubly linked list. The implementation does not use Python's language advantages
 and looks dumb, because it only serves as an data structure exercise and has no
 practical usage.
 """
-from typing import TypeVar, Optional, Sequence
+from typing import TypeVar, Optional, Generic
 from .custom_stack_queue import CustomQueue
-from .. import LinkedListMixin
+from .linked_mixin import LinkedMixin
 from .size_mixin import SizeMixin
 
 
@@ -15,7 +15,7 @@ GT = TypeVar('GT')
 """type: The generic type to represent the element type of the queue."""
 
 
-class LinkedQueue(SizeMixin, CustomQueue[GT]):
+class LinkedQueue(Generic[GT], LinkedMixin[GT], SizeMixin, CustomQueue[GT]):
     """
     `LinkedQueue[T]()` -> a queue based on linked list for values of type `T`.
 
@@ -31,35 +31,7 @@ class LinkedQueue(SizeMixin, CustomQueue[GT]):
         size (int): the current size of the queue
     """
 
-    Node = LinkedListMixin.DoublyNode
-
-    def __init__(self):
-        super().__init__()
-        self.head = None
-        self.tail = None
-
-    def push(self, val: GT) -> None:
-        """Push a value into the end of the queue.
-
-        Args:
-            val: the value to push in
-        """
-        node = self.Node[GT](val)
-        if self.size:
-            self.tail.next = node
-            node.prev = self.tail
-            self.tail = node
-            self.size += 1
-        else:
-            self.head = self.tail = node
-            self.size += 1
-
     def pop(self) -> Optional[GT]:
-        """Pop a value out from the start of the queue.
-
-        Returns:
-            The popped value or `None` if an empty queue
-        """
         if self.size:
             val = self.head.val
             self.head = self.head.next
@@ -70,16 +42,3 @@ class LinkedQueue(SizeMixin, CustomQueue[GT]):
             self.size -= 1
             return val
         return None
-
-    def traverse(self) -> Sequence[GT]:
-        """Traverse all values in the queue and return as a Python `list`.
-
-        Returns:
-            A Python `list` containing all values in the queue
-        """
-        list_ = []
-        node = self.head
-        while node:
-            list_.append(node.val)
-            node = node.next
-        return list_
