+++ data_structures/sequence/stack_queue/linked_stack.py
@@ -5,9 +5,9 @@ doubly linked list. The implementation does not use Python's language advantages
 and looks dumb, because it only serves as an data structure exercise and has no
 practical usage.
 """
-from typing import TypeVar, Optional, Sequence
+from typing import TypeVar, Optional, Generic
 from .custom_stack_queue import CustomStack
-from .. import LinkedListMixin
+from .linked_mixin import LinkedMixin
 from .size_mixin import SizeMixin
 
 
@@ -15,7 +15,7 @@ GT = TypeVar('GT')
 """type: The generic type to represent the element type of the stack."""
 
 
-class LinkedStack(SizeMixin, CustomStack[GT]):
+class LinkedStack(Generic[GT], LinkedMixin[GT], SizeMixin, CustomStack[GT]):
     """
     `LinkedStack[T]()` -> a stack based on linked list for values of type `T`.
 
@@ -31,36 +31,7 @@ class LinkedStack(SizeMixin, CustomStack[GT]):
         size (int): the current size of the stack
     """
 
-    Node = LinkedListMixin.DoublyNode
-    """type: An alias for the correspondent node type."""
-
-    def __init__(self):
-        super().__init__()
-        self.head = None
-        self.tail = None
-
-    def push(self, val: GT) -> None:
-        """Push a value into the open end of the stack.
-
-        Args:
-            val: the value to push in
-        """
-        node = self.Node[GT](val)
-        if self.size:
-            self.tail.next = node
-            node.prev = self.tail
-            self.tail = node
-            self.size += 1
-        else:
-            self.head = self.tail = node
-            self.size += 1
-
     def pop(self) -> Optional[GT]:
-        """Pop a value out from the open end of the stack.
-
-        Returns:
-            The popped value or `None` if an empty stack
-        """
         if self.size:
             val = self.tail.val
             self.tail = self.tail.prev
@@ -71,16 +42,3 @@ class LinkedStack(SizeMixin, CustomStack[GT]):
             self.size -= 1
             return val
         return None
-
-    def traverse(self) -> Sequence[GT]:
-        """Traverse all values in the stack and return as a Python `list`.
-
-        Returns:
-            A Python `list` containing all values in the stack
-        """
-        list_ = []
-        node = self.head
-        while node:
-            list_.append(node.val)
-            node = node.next
-        return list_
